We now have autocomplete on the search field with some mock data. We render a two column content showing ticker and name. We match against both ticker and name.
The table has an autogenerated ID column, and string columns for TICKER and NAME.

We control the query and rendering fully, not using any wizards for those. Next up is multiple selection shown properly, and highlighting of matching parts in the content. Both of these involve javascript changes on the file.

Will also add a command section once the multi verb is complete, so we can select a command list for any entry that starts with a ':'.