As I was debugging I noticed that the header was showing the brand with shadow. It should not have. After chasing this for several hours, I am proud to say that I learned something: make sure you declare your gem versions in the gemfile.
We were using less-rails-bootstrap gem which pulls in Twitter Bootstrap library. In my other machine I used it since July, version 2.0.13 was there already. Unless I ran bundle update, it was OK with the version it found. However this new project pulled in version 2.2.1 of the same gem, and formatting is slightly different.

I locked it to known version since I don't want to spend any more time on styling at this point.